JBIDE-13058: Change versioning pattern for TP
This suggested pattern makes TP and parent pom less independant one
from the other. This should reduce the necessary amount of change
only for versioning.
It also adds the concept of groupId for target platforms, which can
be used to reference jbosstools, jbdevstudio, and other laters (soatools)
Parent pom was updated to use this TP versioning.